+++ api_db.php
@@ -8,9 +8,10 @@
 */
 
 $api = true;
+require 'scripts/pi-hole/php/password.php';
+
 header('Content-type: application/json');
 require 'scripts/pi-hole/php/database.php';
-require 'scripts/pi-hole/php/password.php';
 require 'scripts/pi-hole/php/auth.php';
 require_once 'scripts/pi-hole/php/func.php';
 check_cors();
