+++ style/themes/default-light.css
@@ -158,6 +158,10 @@
   background-color: #3b8ab8;
 }
 
+.login-box {
+  background: #f0f2f4;
+}
+
 .not-used {
   background-color: #eee;
 }
