+++ style/themes/lcars.css
@@ -337,37 +337,33 @@ p.login-box-msg,
   height: 32px;
 }
 
-.form-group.has-error.login-box-msg::before {
+.has-error.login-box-msg::before,
+.has-error.login-box-msg::after {
   content: "ACCESS DENIED";
   margin: -20px -20px 20px;
-  padding: calc(45px + 0.7em) 0 0;
+  padding: calc(176px - 0.5em) 0 0;
   position: absolute;
   inset: 0 0 0 0;
-  color: #630030;
-  font-size: 3.825em;
-  font-size: calc((100vw - var(--sidebar-width) - 60px) / 12);
+  font-size: clamp(40px, calc((100vw - 60px) / 6), 90px);
   line-height: 1;
-  background: rgba(0, 0, 0, 0.88);
+  height: calc(176px + 2em);
+}
+
+.form-group.has-error.login-box-msg::before {
+  color: #600;
+  background-image: linear-gradient(
+    rgba(0, 0, 0, 0),
+    rgba(0, 0, 0, 0.7) 10%,
+    rgba(0, 0, 0, 0.7) 70%,
+    rgba(0, 0, 0, 0)
+  );
+  text-shadow: 0 0 0.25em #000;
 }
 
 .form-group.has-error.login-box-msg::after {
-  content: "ACCESS DENIED";
-  margin: -20px -20px 20px;
-  padding: calc(45px + 0.7em) 0 0;
-  position: absolute;
-  inset: 0 0 0 0;
-  color: #e33100;
-  font-size: 3.825em;
-  font-size: calc((100vw - var(--sidebar-width) - 60px) / 12);
-  line-height: 1;
-  text-shadow: 0 0 5px #e33100;
+  color: #e30;
   animation: 3s infinite Pulse steps(30);
   z-index: 1;
-}
-
-.layout-boxed .form-group.has-error.login-box-msg::before,
-.layout-boxed .form-group.has-error.login-box-msg::after {
-  padding: calc(50% - 0.5em) 0 0;
 }
 
 #loginform .has-error + .row .btn {
@@ -381,10 +377,6 @@ p.login-box-msg,
 #loginform .has-error + .row .btn:active,
 #loginform .has-error + .row .btn:focus {
   background: #f00;
-}
-
-.login-help p {
-  font-size: 13px;
 }
 
 .input-group .input-group-addon {
@@ -1100,6 +1092,11 @@ table.table-bordered.dataTable td {
   background-color: transparent;
 }
 
+.login-box {
+  margin-bottom: 0;
+  background: #181818;
+}
+
 .page-header {
   border: none;
 }
@@ -1552,13 +1549,6 @@ table.dataTable {
 /*** MEDIA QUERIES ***/
 
 /*--- Media - max-width ---*/
-@media (max-width: 991px) {
-  .form-group.has-error.login-box-msg::before,
-  .form-group.has-error.login-box-msg::after {
-    font-size: calc((100vw - var(--sidebar-width) - 60px) / 9);
-  }
-}
-
 @media (max-width: 767px) {
   .sidebar-collapse {
     --sidebar-width: 230px;
@@ -1645,10 +1635,6 @@ table.dataTable {
   .main-footer {
     margin-left: 0;
   }
-  .form-group.has-error.login-box-msg::before,
-  .form-group.has-error.login-box-msg::after {
-    font-size: calc((100vw - 60px) / 6);
-  }
 }
 
 /*--- Media - min-width ---*/
@@ -1716,19 +1702,7 @@ table.dataTable {
   }
 }
 
-@media (min-width: 1320px) {
-  .layout-boxed .form-group.has-error.login-box-msg::before,
-  .layout-boxed .form-group.has-error.login-box-msg::after {
-    font-size: calc((1320px - var(--sidebar-width) - 60px) / 12);
-  }
-}
-
 @media (min-width: 1960px) {
-  .layout-boxed .form-group.has-error.login-box-msg::before,
-  .layout-boxed .form-group.has-error.login-box-msg::after {
-    font-size: calc((1920px - var(--sidebar-width) - 60px) / 12);
-  }
-
   .layout-boxed .wrapper {
     box-shadow: none;
     max-width: 1920px;
